requirements.txt
streamlit
pypdf
python-dotenv
tiktoken
faiss-cpu
sentence-transformers
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq